Pull in `.env` file, auto-restart, and make interactive
Pulls in config from the `.env` file, so that we can store secrets
outside of git if needed.

Turns on auto-restart for data-store, find and submit, so that if they
crash the container reboots automatically. Hopefully more pros than cons
but we could drop this if it ends up annoying.

Opens stdin and enables tty mode, so that eg you can interactively debug
via pdb/ipdb/pdbpp directly if you don't use remote debugging.